+++ CHANGELOG.md
@@ -5,13 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.06] - 2024-12-12
+
+### Added
+
+- Add responsiveness to settings page
+- Add new tailwind classes
+- Add server actions
+
+### Changed
+
+- Change setting page design
+- Change $isFetching store
+
 ## [0.3.05] - 2024-12-11
 
 ### Added
 
 - New InputWithButton component
-- New Popup component
 - New Confirm component
+- New Modal component
 - New Icons
 
 ### Changed